jenkins-plugin/src/main/java/org/eazyportal/plugin/release/jenkins/action/FinalizeSnapshotVersionActionFactory.java
package org.eazyportal.plugin.release.jenkins.action;

import com.google.inject.Inject;
import hudson.Extension;
import hudson.model.InvisibleAction;
import org.eazyportal.plugin.release.core.action.FinalizeSnapshotVersionAction;
import org.eazyportal.plugin.release.jenkins.ReleaseStepConfig;

import java.io.Serializable;

@Extension
public class FinalizeSnapshotVersionActionFactory extends InvisibleAction implements Serializable {

    @Inject
    private transient ReleaseStepConfig releaseStepConfig;

    public FinalizeSnapshotVersionAction create() {
        return new FinalizeSnapshotVersionAction(releaseStepConfig.getScmActions());
    }

}
